+++ gui/main.go
@@ -1,6 +1,16 @@
 package main
 
-import "log"
+import (
+	"embed"
+	"log"
+
+	"github.com/wailsapp/wails/v2"
+	"github.com/wailsapp/wails/v2/pkg/options"
+	"github.com/wailsapp/wails/v2/pkg/options/assetserver"
+)
+
+//go:embed all:frontend/dist
+var assets embed.FS
 
 func main() {
 	if err := run(); err != nil {
@@ -10,6 +20,15 @@ func main() {
 
 func run() error {
 	app := NewApp()
-	_ = app
-	return nil
+	return wails.Run(&options.App{
+		Title:  "Reqlet",
+		Width:  1280,
+		Height: 800,
+		AssetServer: &assetserver.Options{
+			Assets: assets,
+		},
+		OnStartup:  app.startup,
+		OnShutdown: app.shutdown,
+		Bind:       []interface{}{app},
+	})
 }